flask/requirements.txt
Flask
Flask-SQLAlchemy
flask_cors
Flask-Migrate
python-dotenv
psycopg2
pytest
flask-testing
marshmallow
marshmallow_sqlalchemy